backend/requirements.txt
fastapi==0.115.0
starlette>=0.27.0,<1.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
jsonschema==4.23.0
pydash==8.0.5

pytest==8.3.3
pytest-cov==5.0.0
httpx==0.24.1
jsonpath-ng==1.7.0
paho-mqtt==1.6.1
openai==2.6.0
google-auth==2.29.0
requests==2.31.0
python-jose[cryptography]==3.3.0

python-multipart==0.0.9